move channels into upstream
