+++ miner-apps/translator/src/lib/sv2/upstream/mod.rs
@@ -1,4 +1,3 @@
 pub mod common_message_handler;
 pub mod upstream;
 pub use upstream::Upstream;
-pub(super) mod channel;
