+++ miner-apps/translator/src/lib/sv2/upstream/upstream.rs
@@ -2,7 +2,6 @@ use crate::{
     error::{self, TproxyError, TproxyErrorKind, TproxyResult},
     io_task::spawn_io_tasks,
     status::{handle_error, Status, StatusSender},
-    sv2::upstream::channel::UpstreamChannelState,
     utils::UpstreamEntry,
 };
 use async_channel::{unbounded, Receiver, Sender};
@@ -27,6 +26,43 @@ use stratum_apps::{
 use tokio::net::TcpStream;
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, warn};
+
+
+#[derive(Debug, Clone)]
+pub struct UpstreamChannelState {
+    /// Receiver for the SV2 Upstream role
+    pub upstream_receiver: Receiver<Sv2Frame>,
+    /// Sender for the SV2 Upstream role
+    pub upstream_sender: Sender<Sv2Frame>,
+    /// Sender for the ChannelManager to send SV2 frames
+    pub channel_manager_sender: Sender<Sv2Frame>,
+    /// Receiver for the ChannelManager to receive SV2 frames
+    pub channel_manager_receiver: Receiver<Sv2Frame>,
+}
+
+#[cfg_attr(not(test), hotpath::measure_all)]
+impl UpstreamChannelState {
+    pub fn new(
+        upstream_receiver: Receiver<Sv2Frame>,
+        upstream_sender: Sender<Sv2Frame>,
+        channel_manager_sender: Sender<Sv2Frame>,
+        channel_manager_receiver: Receiver<Sv2Frame>,
+    ) -> Self {
+        Self {
+            upstream_receiver,
+            upstream_sender,
+            channel_manager_sender,
+            channel_manager_receiver,
+        }
+    }
+
+    pub fn drop(&self) {
+        debug!("Closing all upstream channels");
+        self.upstream_receiver.close();
+        self.upstream_receiver.close();
+    }
+}
+
 
 /// Manages the upstream SV2 connection to a mining pool or proxy.
 ///
